refactor(reference-data): split account_type_handler.go by operation type
Split the 806-line handler into focused files:
- account_type_query.go: GetDefinition, GetActiveDefinition, ListActive, pagination
- account_type_create.go: CreateDraft, ValidateProductDefinition, CEL/schema validation
- account_type_update.go: UpdateDefinition, ActivateAccountType, DeprecateAccountType
- account_type_handler.go: struct, constructor, error mapping, proto conversions, utilities
