+++ services/reference-data/handler/account_type_handler.go
@@ -1,4 +1,3 @@
-//meridian:large-file - known oversized file; split tracked in backlog
 package handler
 
 import (
@@ -12,7 +11,6 @@ import (
 	"strings"
 
 	"github.com/google/uuid"
-	"github.com/santhosh-tekuri/jsonschema/v5"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 	"google.golang.org/protobuf/types/known/timestamppb"
@@ -57,430 +55,6 @@ func NewAccountTypeService(
 		compiler:           compiler,
 		logger:             logger,
 	}, nil
-}
-
-// GetDefinition retrieves a specific account type definition by ID.
-func (s *AccountTypeService) GetDefinition(ctx context.Context, req *pb.GetDefinitionRequest) (*pb.GetDefinitionResponse, error) {
-	id, err := uuid.Parse(req.GetId())
-	if err != nil {
-		return nil, status.Errorf(codes.InvalidArgument, "invalid id: %v", err)
-	}
-
-	def, err := s.registry.GetDefinitionByID(ctx, id)
-	if err != nil {
-		return nil, s.mapDomainError(ctx, err, "GetDefinition", req.GetId())
-	}
-
-	return &pb.GetDefinitionResponse{
-		Definition: accountTypeToProto(def),
-	}, nil
-}
-
-// GetActiveDefinition retrieves the currently active definition for a given code.
-func (s *AccountTypeService) GetActiveDefinition(ctx context.Context, req *pb.GetActiveDefinitionRequest) (*pb.GetActiveDefinitionResponse, error) {
-	if req.GetCode() == "" {
-		return nil, status.Errorf(codes.InvalidArgument, "code is required")
-	}
-
-	def, err := s.registry.GetActiveDefinition(ctx, req.GetCode())
-	if err != nil {
-		return nil, s.mapDomainError(ctx, err, "GetActiveDefinition", req.GetCode())
-	}
-
-	return &pb.GetActiveDefinitionResponse{
-		Definition: accountTypeToProto(def),
-	}, nil
-}
-
-// ListActive returns all active account type definitions.
-func (s *AccountTypeService) ListActive(ctx context.Context, req *pb.ListActiveRequest) (*pb.ListActiveResponse, error) {
-	defs, err := s.registry.ListActive(ctx)
-	if err != nil {
-		s.logger.Error("failed to list active account types", "error", err)
-		return nil, status.Errorf(codes.Internal, "failed to list active account types: %v", err)
-	}
-
-	defs = filterByBehaviorClass(defs, req.GetBehaviorClassFilter())
-	sort.Slice(defs, func(i, j int) bool {
-		return defs[i].Code < defs[j].Code
-	})
-
-	page, nextPageToken := paginateDefinitions(defs, int(req.GetPageSize()), req.GetPageToken())
-
-	definitions := make([]*pb.AccountTypeDefinition, len(page))
-	for i, def := range page {
-		definitions[i] = accountTypeToProto(def)
-	}
-
-	return &pb.ListActiveResponse{
-		Definitions:   definitions,
-		NextPageToken: nextPageToken,
-	}, nil
-}
-
-func filterByBehaviorClass(defs []*accounttype.Definition, filter pb.BehaviorClass) []*accounttype.Definition {
-	if filter == pb.BehaviorClass_BEHAVIOR_CLASS_UNSPECIFIED {
-		return defs
-	}
-	domainBC := protoBehaviorClassToDomain(filter)
-	var filtered []*accounttype.Definition
-	for _, def := range defs {
-		if def.BehaviorClass == domainBC {
-			filtered = append(filtered, def)
-		}
-	}
-	return filtered
-}
-
-func paginateDefinitions(defs []*accounttype.Definition, reqPageSize int, pageToken string) ([]*accounttype.Definition, string) {
-	pageSize := normalizeAccountTypePageSize(reqPageSize)
-	startIdx := findStartIndex(defs, pageToken)
-
-	if startIdx >= len(defs) {
-		return nil, ""
-	}
-
-	end := startIdx + pageSize
-	if end > len(defs) {
-		end = len(defs)
-	}
-	page := defs[startIdx:end]
-
-	var nextPageToken string
-	if end < len(defs) {
-		nextPageToken = page[len(page)-1].Code
-	}
-	return page, nextPageToken
-}
-
-func findStartIndex(defs []*accounttype.Definition, pageToken string) int {
-	if pageToken == "" {
-		return 0
-	}
-	for i, def := range defs {
-		if def.Code > pageToken {
-			return i
-		}
-	}
-	return len(defs)
-}
-
-func normalizeAccountTypePageSize(pageSize int) int {
-	if pageSize <= 0 {
-		return DefaultPageSize
-	}
-	if pageSize > MaxPageSize {
-		return MaxPageSize
-	}
-	return pageSize
-}
-
-// CreateDraft creates a new account type definition in DRAFT status.
-func (s *AccountTypeService) CreateDraft(ctx context.Context, req *pb.CreateDraftRequest) (*pb.CreateDraftResponse, error) {
-	defaultConversionMethodID, defaultConversionMethodVersion, err := parseConversionMethodPair(
-		req.GetDefaultConversionMethodId(), req.GetDefaultConversionMethodVersion())
-	if err != nil {
-		return nil, err
-	}
-
-	def, err := accounttype.NewDefinition(accounttype.NewDefinitionParams{
-		Code:                           req.GetCode(),
-		DisplayName:                    req.GetDisplayName(),
-		Description:                    req.GetDescription(),
-		NormalBalance:                  protoBehaviorNormalBalanceToDomainString(req.GetNormalBalance()),
-		BehaviorClass:                  protoBehaviorClassToDomainString(req.GetBehaviorClass()),
-		InstrumentCode:                 req.GetInstrumentCode(),
-		DefaultSagaPrefix:              req.GetDefaultSagaPrefix(),
-		DefaultConversionMethodID:      defaultConversionMethodID,
-		DefaultConversionMethodVersion: defaultConversionMethodVersion,
-		ValidationCEL:                  req.GetValidationCel(),
-		BucketingCEL:                   req.GetBucketingCel(),
-		EligibilityCEL:                 req.GetEligibilityCel(),
-		AttributeSchema:                toRawJSON(req.GetAttributeSchema()),
-		Attributes:                     stringMapToAnyMap(req.GetAttributes()),
-		Compiler:                       s.compiler,
-	})
-	if err != nil {
-		return nil, s.mapDomainError(ctx, err, "CreateDraft", req.GetCode())
-	}
-
-	if err := s.registry.CreateDraft(ctx, def); err != nil {
-		return nil, s.mapDomainError(ctx, err, "CreateDraft", req.GetCode())
-	}
-
-	s.logger.Info("account type draft created",
-		"code", def.Code,
-		"version", def.Version)
-
-	return &pb.CreateDraftResponse{
-		Definition: accountTypeToProto(def),
-	}, nil
-}
-
-// UpdateDefinition modifies a DRAFT account type definition.
-func (s *AccountTypeService) UpdateDefinition(ctx context.Context, req *pb.UpdateDefinitionRequest) (*pb.UpdateDefinitionResponse, error) {
-	id, err := uuid.Parse(req.GetId())
-	if err != nil {
-		return nil, status.Errorf(codes.InvalidArgument, "invalid id: %v", err)
-	}
-
-	existing, err := s.registry.GetDefinitionByID(ctx, id)
-	if err != nil {
-		return nil, s.mapDomainError(ctx, err, "UpdateDefinition", req.GetId())
-	}
-
-	defaultConversionMethodID, defaultConversionMethodVersion, parseErr := parseConversionMethodPair(
-		req.GetDefaultConversionMethodId(), req.GetDefaultConversionMethodVersion())
-	if parseErr != nil {
-		return nil, parseErr
-	}
-
-	updates := &accounttype.Definition{
-		DisplayName:                    req.GetDisplayName(),
-		Description:                    req.GetDescription(),
-		InstrumentCode:                 req.GetInstrumentCode(),
-		DefaultSagaPrefix:              req.GetDefaultSagaPrefix(),
-		DefaultConversionMethodID:      defaultConversionMethodID,
-		DefaultConversionMethodVersion: defaultConversionMethodVersion,
-		ValidationCEL:                  req.GetValidationCel(),
-		BucketingCEL:                   req.GetBucketingCel(),
-		EligibilityCEL:                 req.GetEligibilityCel(),
-		AttributeSchema:                toRawJSON(req.GetAttributeSchema()),
-		Attributes:                     stringMapToAnyMap(req.GetAttributes()),
-	}
-
-	if err := s.registry.UpdateDefinition(ctx, existing.Code, existing.Version, updates); err != nil {
-		return nil, s.mapDomainError(ctx, err, "UpdateDefinition", req.GetId())
-	}
-
-	updated, err := s.registry.GetDefinitionByID(ctx, id)
-	if err != nil {
-		return nil, s.mapDomainError(ctx, err, "UpdateDefinition", req.GetId())
-	}
-
-	s.logger.Info("account type updated",
-		"code", existing.Code,
-		"version", existing.Version)
-
-	return &pb.UpdateDefinitionResponse{
-		Definition: accountTypeToProto(updated),
-	}, nil
-}
-
-// ActivateAccountType transitions an account type from DRAFT to ACTIVE.
-func (s *AccountTypeService) ActivateAccountType(ctx context.Context, req *pb.ActivateAccountTypeRequest) (*pb.ActivateAccountTypeResponse, error) {
-	id, err := uuid.Parse(req.GetId())
-	if err != nil {
-		return nil, status.Errorf(codes.InvalidArgument, "invalid id: %v", err)
-	}
-
-	existing, err := s.registry.GetDefinitionByID(ctx, id)
-	if err != nil {
-		return nil, s.mapDomainError(ctx, err, "ActivateAccountType", req.GetId())
-	}
-
-	if err := s.registry.ActivateAccountType(ctx, existing.Code, existing.Version); err != nil {
-		return nil, s.mapDomainError(ctx, err, "ActivateAccountType", req.GetId())
-	}
-
-	activated, err := s.registry.GetDefinitionByID(ctx, id)
-	if err != nil {
-		return nil, s.mapDomainError(ctx, err, "ActivateAccountType", req.GetId())
-	}
-
-	s.logger.Info("account type activated",
-		"code", existing.Code,
-		"version", existing.Version)
-
-	return &pb.ActivateAccountTypeResponse{
-		Definition: accountTypeToProto(activated),
-	}, nil
-}
-
-// DeprecateAccountType transitions an account type from ACTIVE to DEPRECATED.
-func (s *AccountTypeService) DeprecateAccountType(ctx context.Context, req *pb.DeprecateAccountTypeRequest) (*pb.DeprecateAccountTypeResponse, error) {
-	id, err := uuid.Parse(req.GetId())
-	if err != nil {
-		return nil, status.Errorf(codes.InvalidArgument, "invalid id: %v", err)
-	}
-
-	existing, err := s.registry.GetDefinitionByID(ctx, id)
-	if err != nil {
-		return nil, s.mapDomainError(ctx, err, "DeprecateAccountType", req.GetId())
-	}
-
-	var successorID *uuid.UUID
-	if req.GetSuccessorId() != "" {
-		parsed, parseErr := uuid.Parse(req.GetSuccessorId())
-		if parseErr != nil {
-			return nil, status.Errorf(codes.InvalidArgument, "invalid successor_id: %v", parseErr)
-		}
-		successorID = &parsed
-	}
-
-	if err := s.registry.DeprecateAccountType(ctx, existing.Code, existing.Version, successorID); err != nil {
-		return nil, s.mapDomainError(ctx, err, "DeprecateAccountType", req.GetId())
-	}
-
-	deprecated, err := s.registry.GetDefinitionByID(ctx, id)
-	if err != nil {
-		return nil, s.mapDomainError(ctx, err, "DeprecateAccountType", req.GetId())
-	}
-
-	s.logger.Info("account type deprecated",
-		"code", existing.Code,
-		"version", existing.Version)
-
-	return &pb.DeprecateAccountTypeResponse{
-		Definition: accountTypeToProto(deprecated),
-	}, nil
-}
-
-// ValidateProductDefinition validates an account type definition without persisting it.
-func (s *AccountTypeService) ValidateProductDefinition(ctx context.Context, req *pb.ValidateProductDefinitionRequest) (*pb.ValidateProductDefinitionResponse, error) {
-	def := req.GetDefinition()
-	if def == nil {
-		return nil, status.Errorf(codes.InvalidArgument, "definition is required")
-	}
-
-	validationErrors := make([]*pb.ValidationError, 0, 4)
-
-	validationErrors = append(validationErrors, s.validateCELExpressions(def)...)
-	schemaValid, schemaErrs := validateAttributeSchema(def.AttributeSchema)
-	validationErrors = append(validationErrors, schemaErrs...)
-	validationErrors = append(validationErrors, validateAttrsAgainstSchema(def, schemaValid)...)
-	validationErrors = append(validationErrors, s.validateInstrumentRefs(ctx, def)...)
-
-	return &pb.ValidateProductDefinitionResponse{
-		Valid:  len(validationErrors) == 0,
-		Errors: validationErrors,
-	}, nil
-}
-
-func (s *AccountTypeService) validateCELExpressions(def *pb.AccountTypeDefinition) []*pb.ValidationError {
-	var errs []*pb.ValidationError
-	if def.ValidationCel != "" {
-		if err := s.compiler.ValidateValidationCEL(def.ValidationCel); err != nil {
-			errs = append(errs, celCompileError("validation_cel", err))
-		}
-	}
-	if def.BucketingCel != "" {
-		if err := s.compiler.ValidateBucketingCEL(def.BucketingCel); err != nil {
-			errs = append(errs, celCompileError("bucketing_cel", err))
-		}
-	}
-	if def.EligibilityCel != "" {
-		if err := s.compiler.ValidateEligibilityCEL(def.EligibilityCel); err != nil {
-			errs = append(errs, celCompileError("eligibility_cel", err))
-		}
-	}
-	return errs
-}
-
-func validateAttributeSchema(schema string) (bool, []*pb.ValidationError) {
-	if schema == "" {
-		return false, nil
-	}
-	if !json.Valid([]byte(schema)) {
-		return false, []*pb.ValidationError{{
-			Field:     "attribute_schema",
-			ErrorCode: "INVALID_JSON",
-			Message:   "attribute_schema is not valid JSON",
-		}}
-	}
-	c := jsonschema.NewCompiler()
-	if err := c.AddResource("schema.json", strings.NewReader(schema)); err != nil {
-		return false, []*pb.ValidationError{{
-			Field:     "attribute_schema",
-			ErrorCode: "INVALID_JSON_SCHEMA",
-			Message:   fmt.Sprintf("invalid JSON Schema: %v", err),
-		}}
-	}
-	if _, err := c.Compile("schema.json"); err != nil {
-		return false, []*pb.ValidationError{{
-			Field:     "attribute_schema",
-			ErrorCode: "INVALID_JSON_SCHEMA",
-			Message:   fmt.Sprintf("JSON Schema compilation failed: %v", err),
-		}}
-	}
-	return true, nil
-}
-
-func validateAttrsAgainstSchema(def *pb.AccountTypeDefinition, schemaValid bool) []*pb.ValidationError {
-	if !schemaValid || len(def.Attributes) == 0 {
-		return nil
-	}
-	attrs := stringMapToAnyMap(def.Attributes)
-	c := jsonschema.NewCompiler()
-	_ = c.AddResource("schema.json", strings.NewReader(def.AttributeSchema))
-	compiled, _ := c.Compile("schema.json")
-	if err := compiled.Validate(attrs); err != nil {
-		return []*pb.ValidationError{{
-			Field:     "attributes",
-			ErrorCode: "SCHEMA_VALIDATION_FAILED",
-			Message:   fmt.Sprintf("attributes do not validate against schema: %v", err),
-		}}
-	}
-	return nil
-}
-
-func (s *AccountTypeService) validateInstrumentRefs(ctx context.Context, def *pb.AccountTypeDefinition) []*pb.ValidationError {
-	if s.instrumentRegistry == nil {
-		return nil
-	}
-	var errs []*pb.ValidationError
-
-	if def.InstrumentCode != "" {
-		if _, err := s.instrumentRegistry.GetActiveDefinition(ctx, def.InstrumentCode); err != nil {
-			ve := &pb.ValidationError{
-				Field:       "instrument_code",
-				ErrorCode:   "UNRESOLVABLE_REFERENCE",
-				Message:     fmt.Sprintf("instrument %q not found or not ACTIVE", def.InstrumentCode),
-				Suggestions: s.instrumentSuggestions(ctx, def.InstrumentCode),
-			}
-			errs = append(errs, ve)
-		}
-	}
-
-	for i, vmt := range def.ValuationMethods {
-		if vmt.InputInstrument != "" {
-			if _, err := s.instrumentRegistry.GetActiveDefinition(ctx, vmt.InputInstrument); err != nil {
-				ve := &pb.ValidationError{
-					Field:       fmt.Sprintf("valuation_methods[%d].input_instrument", i),
-					ErrorCode:   "UNRESOLVABLE_REFERENCE",
-					Message:     fmt.Sprintf("instrument %q not found or not ACTIVE", vmt.InputInstrument),
-					Suggestions: s.instrumentSuggestions(ctx, vmt.InputInstrument),
-				}
-				errs = append(errs, ve)
-			}
-		}
-	}
-
-	return errs
-}
-
-// instrumentSuggestions returns Levenshtein-based suggestions for a typo instrument code.
-func (s *AccountTypeService) instrumentSuggestions(ctx context.Context, code string) []string {
-	if s.instrumentRegistry == nil {
-		return nil
-	}
-	actives, err := s.instrumentRegistry.ListActive(ctx)
-	if err != nil {
-		return nil
-	}
-	candidates := make([]string, len(actives))
-	for i, a := range actives {
-		candidates[i] = a.Code
-	}
-	return findSuggestions(code, candidates)
-}
-
-func celCompileError(field string, err error) *pb.ValidationError {
-	return &pb.ValidationError{
-		Field:     field,
-		ErrorCode: "CEL_COMPILE_ERROR",
-		Message:   err.Error(),
-	}
 }
 
 // accountTypeErrorMapping maps domain errors to gRPC status codes.
